refactor(pty): relocate PtyBridge from mux/ to src/pty/ leaf
Pure move so web-ui/ can consume the PTY core without a web-ui->mux
import edge. git mv + 3 import-site fixes; no behavior change. tsc,
pty-bridge unit test, eslint, prettier all green.
